Extract temporary MongoDB endpoint helper
