+++ instrumentation/mongo/mongo-3.1/library/src/main/java/io/opentelemetry/instrumentation/mongo/v3_1/internal/MongoServerTarget.java
@@ -61,7 +61,7 @@ public class MongoServerTarget {
         if (unixSocketTarget == null) {
           return null;
         }
-      } else if (hasUnsafeEncodedIpv6Zone(host)) {
+      } else if (MongoServerEndpointUtil.hasUnsafeEncodedIpv6Zone(host)) {
         return null;
       }
       Integer port = unixSocket ? null : seed.getPort();
@@ -130,34 +130,6 @@ public class MongoServerTarget {
       return host.substring(1, host.length() - 1);
     }
     return host;
-  }
-
-  // Temporary helper copied from PR #20015.
-  // Remove it after PR #20015 merges and use DbServerEndpointUtil instead.
-  // Migration tracked in issue #20016.
-  private static boolean hasUnsafeEncodedIpv6Zone(String host) {
-    int zoneSeparator = host.indexOf('%');
-    return zoneSeparator >= 0 && startsWithEncodedDelimiter(host.substring(zoneSeparator + 1));
-  }
-
-  private static boolean startsWithEncodedDelimiter(String value) {
-    if (value.length() < 2) {
-      return false;
-    }
-    int high = Character.digit(value.charAt(0), 16);
-    int low = Character.digit(value.charAt(1), 16);
-    if (high < 0 || low < 0) {
-      return false;
-    }
-    char decoded = (char) ((high << 4) + low);
-    return decoded == ':'
-        || decoded == '@'
-        || decoded == '/'
-        || decoded == '?'
-        || decoded == '#'
-        || decoded == '\\'
-        || decoded == '%'
-        || decoded == '=';
   }
 
   private static boolean isUnixSocket(ServerAddress seed, String host) {
